requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
oauthlib==3.2.0
python-telegram-bot==13.13
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tornado==6.2
tweepy==4.10.1
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
